Express Configuration
General tidy up of the express configuration code to reduce repetition, increase readability and make subsequent refactoring easier
